models: Add personal_recipient foreign key in UserProfile.
This is adds a foreign key to the corresponding Recipient for PERSONAL
messages for each user.

In the migration for setting the field correctly for existing users, we
do a direct SQL query (because Django doesn't provide any good method
for doing it properly in bulk using the ORM.). To avoid caching-related
bugs with cached UserProfile instances not having this field
synchronised with the database, we flush all user_profile caches in that
migration.

A consequence of this change to the model is that a bit of code needs to
be added to the functions responsible for creating new users (to set the
field after the Recipient object gets created).

Also an adjustment is needed in the import system - this introduces a
circular relation between Recipient and UserProfile. The field cannot be
set until the Recipient objects have been created, but UserProfiles need
to be created before their corresponding Recipients. We deal with this
by first importing UserProfiles same way as before, but we leave the
personal_recipient field uninitialized. After creating the Recipient
objects, we call a function to set the field for all the imported users
in bulk.